+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Changed
 * Command line arguments are parsed with tools.cli now.
 
+### Added
+* Option to define the default reader fn from the CLI arguments.
+
 ## [0.2.0] - 2018-10-05
 ### Changed
 * Replaced launch script in favor of native image built with GraalVM.